fix(move-session): move test out of auto-loaded extensions and extract cwd helper
