+++ extensions/move-session.ts
@@ -28,27 +28,12 @@ import {
     renameSync,
     unlinkSync,
 } from "node:fs";
-import { resolve } from "node:path";
+
+import { normalizeTargetCwd } from "./move-session-lib/normalize-target-cwd";
 
 const TRASH_TIMEOUT_MS = 5000;
 const HEADER_READ_MAX = 8192;
 const COPY_CHUNK_SIZE = 65_536;
-
-export function normalizeTargetCwd(
-    rawTargetCwd: string,
-    env: NodeJS.ProcessEnv = process.env,
-    cwd: string = process.cwd(),
-): string {
-    let targetCwd = rawTargetCwd;
-    if (/^~(?=$|\/)/.test(rawTargetCwd)) {
-        const home = env.HOME || env.USERPROFILE;
-        if (!home) {
-            throw new Error("Cannot expand '~': $HOME is not set");
-        }
-        targetCwd = rawTargetCwd.replace(/^~(?=$|\/)/, home);
-    }
-    return resolve(cwd, targetCwd);
-}
 
 /**
  * Remove parentSession from the first JSONL header line without loading
